requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==4.2.11
django-jazzmin==2.6.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
